+++ tools/pfe-tools/dev-server.ts
@@ -1,21 +1,32 @@
 import type { Plugin } from '@web/dev-server-core';
 import type { DevServerConfig } from '@web/dev-server';
 import type { InjectSetting } from '@web/dev-server-import-maps/dist/importMapsPlugin';
-import type { Context, Middleware, Next } from 'koa';
+import type { Context, Next } from 'koa';
+import type { LitCSSOptions } from 'web-dev-server-plugin-lit-css';
 
-import { readdir, stat } from 'fs/promises';
+import 'urlpattern-polyfill';
+
+import { readdir, readFile, stat } from 'fs/promises';
 import { join, dirname } from 'path';
 import { fileURLToPath } from 'url';
 
-import litcssRollup from 'rollup-plugin-lit-css';
 import rollupReplace from '@rollup/plugin-replace';
 import nunjucks from 'nunjucks';
+import _glob from 'glob';
 
+import { litCss } from 'web-dev-server-plugin-lit-css';
 import { fromRollup } from '@web/dev-server-rollup';
 import { esbuildPlugin } from '@web/dev-server-esbuild';
 import { importMapsPlugin } from '@web/dev-server-import-maps';
 import { transformSass } from './esbuild.js';
 import { createRequire } from 'module';
+import { promisify } from 'util';
+
+const glob = promisify(_glob);
+const require = createRequire(import.meta.url);
+const exists = (x: string) => stat(x).then(() => true, () => false);
+const replace = fromRollup(rollupReplace);
+const env = nunjucks.configure(fileURLToPath(new URL('./demo', import.meta.url)));
 
 export interface PfeDevServerConfigOptions extends DevServerConfig {
   /** Extra dev server plugins */
@@ -23,19 +34,16 @@ export interface PfeDevServerConfigOptions extends DevServerConfig {
   plugins?: Plugin[];
   importMap?: InjectSetting['importMap'];
   hostname?: string;
+  watchFiles: string;
+  litcssOptions: LitCSSOptions,
   site?: {
+    tagPrefix: `${string}-`;
     title: string;
     logoUrl: string;
     githubUrl: string;
     description: string;
   };
 }
-
-const require = createRequire(import.meta.url);
-const exists = (x: string) => stat(x).then(() => true, () => false);
-const litcss = fromRollup(litcssRollup);
-const replace = fromRollup(rollupReplace);
-const env = nunjucks.configure(fileURLToPath(new URL('./demo', import.meta.url)));
 
 const SITE_DEFAULTS = {
   title: 'PatternFly Elements',
@@ -44,54 +52,18 @@ const SITE_DEFAULTS = {
   description: 'PatternFly Elements: A set of community-created web components based on PatternFly design.',
 };
 
-function appendLines(body: string, ...lines: string[]): string {
-  return [body, ...lines].join('\n');
+/** Prettify a tag name, stripping the prefix and capitalizing the rest */
+function prettyTag(tagName: string, prefix: `${string}-` = 'pfe-'): string {
+  return tagName
+    .replace(prefix, '')
+    .toLowerCase()
+    .replace(/(?:^|[\s-/])\w/g, x => x.toUpperCase())
+    .replace(/-/g, ' ');
 }
 
-async function tryRead(pathname: string) {
-  return readdir(pathname).catch(() => []);
-}
-
-/**
- * Binds data from the node context to the dev-server's browser context.
- * exposed data is available by importing `@patternfly/pfe-tools/environment.js`.
- */
-function bindNodeDataToBrowser(options?: PfeDevServerConfigOptions): Plugin {
-  const { rootDir = process.cwd() } = options ?? {};
-  return {
-    name: 'bind-node-data-to-browser',
-    async transform(context) {
-      if (context.path.match(/pfe-tools\/environment\.(?:j|t)s$/)) {
-        // TODO: calculate export names from npm workspaces
-        //       for now, `elements` is the only conventionally-supported workspace
-        const elements = await tryRead(join(rootDir, 'elements'));
-        const core = await tryRead(join(rootDir, 'core'));
-        const transformed = appendLines(
-          context.body as string,
-          `export const elements = ${JSON.stringify(elements)};`,
-          `export const core     = ${JSON.stringify(core)};`
-        );
-        return transformed;
-      }
-    },
-  };
-}
-
-function scssMimeType(): Plugin {
-  return {
-    name: 'scss-mime-type',
-    resolveMimeType(context) {
-    // ensure .scss files are loaded by the browser as javascript
-      if (context.path.endsWith('.scss')) {
-        return 'js';
-      }
-    },
-  };
-}
-
+/** Ugly, ugly hack to resolve packages from the local monorepo */
 function tryToResolve(source: string, context: import('koa').Context) {
   let resolved = '';
-
   try {
     resolved = require.resolve(source);
   } catch (error) {
@@ -119,8 +91,8 @@ function tryToResolve(source: string, context: import('koa').Context) {
 /**
  * Resolves local monorepo package imports. Needed because we consume our own monorepo packages
  */
-export function resolveLocalFilesFromTypeScriptSources(options: PfeDevServerConfigOptions): Plugin {
-  const rootDir = options.rootDir as string;
+export function resolveLocalFilesFromTypeScriptSources(options: Partial<PfeDevServerConfigOptions> & { rootDir: string }): Plugin {
+  const { rootDir } = options;
   return {
     name: 'resolve-local-monorepo-packages-from-ts-sources',
     async transformImport({ source, context }) {
@@ -143,30 +115,72 @@ export function resolveLocalFilesFromTypeScriptSources(options: PfeDevServerConf
   };
 }
 
-function nunjucksSPAMiddleware(options?: PfeDevServerConfigOptions): Middleware {
-  const model = { ...SITE_DEFAULTS, ...options?.site };
-  return function(ctx, next) {
-    if (!(ctx.method !== 'HEAD' && ctx.method !== 'GET' || ctx.path.includes('.'))) {
-      const transformed = env.render('index.html', model);
-      ctx.body = transformed;
-      ctx.type = 'html';
-      ctx.status = 200;
+/**
+ * Renders the demo page for a given url
+ */
+async function renderURL(ctx: Context, env: nunjucks.Environment, pattern: URLPattern, options?: PfeDevServerConfigOptions): Promise<string> {
+  const url = new URL(ctx.request.url, `http://${ctx.request.headers.host}`);
+  const { rootDir = process.cwd() } = options ?? {};
+  const { element } = pattern.exec(url)?.pathname?.groups ?? {};
+  const base = element?.match(/^pfe-(core|sass|styles)$/) ? 'core' : 'elements';
+  const basePath = element && join(rootDir, base, element, 'demo');
+  const demoPath = basePath && join(basePath, `${element}.html`);
+  return env.render('index.njk', {
+    ...SITE_DEFAULTS,
+    ...options?.site,
+    core: await readdir(join(rootDir, 'core')).catch(() => []),
+    context: ctx,
+    demo: demoPath && await readFile(demoPath, 'utf-8'),
+    element,
+    elements: await readdir(join(rootDir, 'elements')).catch(() => []),
+    manifest: basePath && (await exists(join(basePath, '..', 'custom-elements.json')) ? `/${base}/${element}/custom-elements.json` : '/custom-elements.json'),
+    title: element ? `${prettyTag(element, options?.site?.tagPrefix)} | ${options?.site?.title ?? 'PatternFly Elements'}` : (options?.site?.title ?? SITE_DEFAULTS.title),
+    script: basePath && await readFile(join(basePath, `${element}.js`), 'utf-8'),
+  });
+}
+
+/**
+ * Generate HTML for each component by adding it's demo to a
+ * Declarative Shadow DOM template.
+ * @see https://web.dev/declarative-shadow-dom/
+ * Watch repository source files and reload the page when they change
+ */
+function pfeDevServerPlugin(options?: PfeDevServerConfigOptions): Plugin {
+  const pattern = new URLPattern({ pathname: '/demo/:element/' });
+  const env = nunjucks.configure(fileURLToPath(new URL('./demo', import.meta.url)));
+  env.addFilter('prettyTag', x => prettyTag(x, options?.site?.tagPrefix));
+
+  return {
+    name: 'pfe-dev-server',
+    async serverStart({ fileWatcher, app }) {
+      /** Render the demo page */
+      app.use(async function nunjucksMiddleware(ctx, next) {
+        if ((options?.loadDemo ?? true) && !(ctx.method !== 'HEAD' && ctx.method !== 'GET' || ctx.path.includes('.'))) {
+          ctx.type = 'html';
+          ctx.status = 200;
+          ctx.body = await renderURL(ctx, env, pattern, options);
+        }
+        return next();
+      });
+
+      const files = await glob(options?.watchFiles ?? '{elements,core}/**/*.{ts,js,css,scss,html}', { cwd: process.cwd() });
+      for (const file of files) {
+        fileWatcher.add(file);
+      }
     }
-    return next();
   };
 }
 
-function cors(context: Context, next: Next) {
-  context.set('Access-Control-Allow-Origin', '*');
+/** CORS middleware */
+function cors(ctx: Context, next: Next) {
+  ctx.set('Access-Control-Allow-Origin', '*');
   return next();
 }
 
 /**
  * Creates a default config for PFE's dev server.
  */
-export function pfeDevServerConfig(_options?: PfeDevServerConfigOptions): DevServerConfig {
-  const { importMap, site, loadDemo = true, ...options } = _options ?? {};
-
+export function pfeDevServerConfig(options?: PfeDevServerConfigOptions): DevServerConfig {
   /**
    * Plain case: this file is running from `/node_modules/@patternfly/pfe-tools`.
    *             two dirs up from here is `node_modules`, so we just shear it clean off the path string
@@ -188,16 +202,15 @@ export function pfeDevServerConfig(_options?: PfeDevServerConfigOptions): DevSer
     ...options ?? {},
 
     middleware: [
-      ...loadDemo ? [
-        nunjucksSPAMiddleware(_options),
-      ] : [],
       cors,
       ...options?.middleware ?? [],
     ],
 
     plugins: [
       ...options?.plugins ?? [],
-      ...(importMap) ? [importMapsPlugin({ inject: { importMap } })] : [],
+
+      ...options?.importMap ? [importMapsPlugin({ inject: { importMap: options.importMap } })] : [],
+
       // ordinarily, the packages' export conditions specify to
       //   1. import the root from a '.js'
       //   2. append '.js' to any subpath imports
@@ -208,21 +221,24 @@ export function pfeDevServerConfig(_options?: PfeDevServerConfigOptions): DevSer
       // we hope it will work in most cases. If you have a problem loading from packages in the dev
       // server, please open an issue at https://github.com/patternfly/patternfly-elements/issues/new/
       resolveLocalFilesFromTypeScriptSources({ ...options, rootDir }),
-      // importMapsPlugin({ inject: { importMap: { imports } } }),
+
       // serve typescript sources as javascript
-      esbuildPlugin({ ts: true }),
-      // bind data from node to the browser
-      // e.g. the file listing of the `elements/` dir
-      // so that we can list the elements by name
-      bindNodeDataToBrowser({ rootDir }),
+      esbuildPlugin({
+        ts: true,
+        // see https://github.com/evanw/esbuild/issues/2220
+        target: 'es2020'
+      }),
+
       // load .scss files as lit CSSResult modules
-      litcss({ include: ['**/*.scss'], transform: transformSass }),
+      litCss(options?.litcssOptions ?? { include: /\.scss$/, transform: transformSass }),
+
       replace({
         'preventAssignment': true,
         'process.env.NODE_ENV': JSON.stringify('production'),
       }),
-      // Ensure .scss files are loaded as js modules, as `litcss` plugin transforms them to such
-      scssMimeType(),
+
+      // Dev server app which loads component demo files
+      pfeDevServerPlugin(options),
     ],
   };
 }